Added math functions for fibonacci and factorial and an attempt to make them print out
I am attempting to make the two functions print to the first 15 digits, but am currently unsuccessful (mainly due to my impatience). Even at an interval of 5 seconds, my computer turns itself into an airplane and begins to take off as I try to run the math for each prime; scaring the hell out of me even after I waited a half hour and was only able to obtain the fibonacci for the multiprocessing prime. Even though I am still unable to fully obtain results, I'm fairly certain this is a decently correct method of doing the math (or at least one that I can still understand as I write it) after doing an insane amount of research into how to optimize the code.