Added integer bypass and now it workssssss
My 5 second interval testing finally returned some numbers! Finally waited long enough for the program to fully run, got an integer error for the max python integers in my math strings. Found the way to bypass it with the int_max set to an arbitrarily large number and it actually gave me an output lets goooo