New lab 11 workfile so it could be done in jupyter
having a crazy time with this one. Can't seem to clean the data properly and have been trying on and off the past couple days. Converted my work to jupyter so i could test the individual blocks faster and am currently able to plot the raw data, just not my "cleaned" version. Assuming that it's because i am not converting the strings properly so they are being dropped which leads to the columns being empty since i continue to get the "no data in column" message