Create downloader.py
downloader combines the parsing and downloading into one file, might break them up later. Parses XML and turns it into JSON, then uses threading to download it. I also added a print for the exchange rates for the target currency against the others for each day in the data as a way to check the code was properly downloading/reading the files which actually worked a lot better than I thought it would (also took forever to get right because I kept misusing the things it would look for to print)