Update downloader.py
Also updated some parts of the downloader to lighten the load and allow me to retry the xml work if they fail as they have been during testing. Added a delay to the threading to stop it from going over. And am currently considering moving/deleting the echange_rate_print as it doesn't do much outside of confirming things were working for part 1