Update downloader.py
Added code to get json data together for EDA. The goal is to use the five main currencies (USD, EUR, GBP, JPY, CNY) from the instructions and see how they trended over the full period of time to start. I got the exchange rates and volatility of the 5 compared to one another using each as a base which will be expanded upon if i can manage it. Currently JPY is the outlier as the yen has a super high exchange rate in comparison to the others just out the gate. Going to add the jupyter file and hopefully my current outputs are also shown. I want to try and get some other graphs and maybe change to a lower value currency other than JPY so that the graphs can be read and analyzed better by me