Added jupyter file of EDA
Focused for now on the exchange rate and volatility between the 5 currencies. Took a long time (almost the full day) to get working to a point where I could upload it as I accidentally had the files being downloaded to the wrong directory and also an issue with the way I was trying to read the files so the graphs would initially show up as empty. Used the logs I implemented throughout my code to eventually figure out what was happening and got it fixed though. I plan to try and get some other graphs and data, but I feel like this is a good place to start the analysis of the 5 currencies. At the point of the last commit, I was pretty much giving up and ready to throw in the towel, just uploading what was actually correctly implemented until I finally figured out my issues. From there it has been a trip trying to fix what I had messed up and finally getting some actual data back in my graphs rather than just empty plots and error codes.